Moved validator to utilities and fixed progress bar bug
