+++ SplitBill/ViewModel/SharedData.swift
@@ -123,37 +123,6 @@ final class SharedData: ObservableObject {
         }
     }
     
-    func formatBillAmount(_ imput: String) -> String {
-        if imput.count > 10 {
-            return String(imput.prefix(10))
-        }
-        
-        let decimalSeparator = Locale.current.decimalSeparator ?? "."
-        let allowedCharacters = CharacterSet(charactersIn: "0123456789" + decimalSeparator)
-        
-        let filtered = imput.filter { char in
-            let stringChar = String(char)
-            return stringChar.rangeOfCharacter(from: allowedCharacters) != nil
-        }
-        
-        let normalized = filtered.replacingOccurrences(of: decimalSeparator == "." ? "," : ".", with: decimalSeparator)
-        let components = normalized.components(separatedBy: decimalSeparator)
-        
-        if components.count > 2 {
-            let firstPart = components[0]
-            let remaining = components[1...].joined()
-            return firstPart + decimalSeparator + remaining
-        }
-        
-        if components.count == 2 {
-            let integerPart = components[0]
-            let fractionalPart = String(components[1].prefix(2))
-            return integerPart + decimalSeparator + fractionalPart
-        }
-        
-        return normalized
-    }
-    
     func resetToInitialState() {
         participants = []
         billAmount = ""
